+++ src/app/layout.tsx
@@ -3,8 +3,8 @@ import "@mantine/core/styles.css";
 
 import { MantineProvider } from "@mantine/core";
 import { McMurtryTheme } from "./theme";
-import { Header } from "../components/Header";
-import { Footer } from "../components/Footer";
+import { Header } from "../components/header/Header";
+import { Footer } from "../components/footer/Footer";
 
 export const metadata: Metadata = {
   title: "McMurtry College",
